requirements.txt
boto3==1.17.91
botocore==1.20.91
Django==3.2.4
djangorestframework==3.12.4